4.0 基本完成

+++ GripBall/GripBall/view/Practice/JCManChart.h
@@ -9,6 +9,6 @@
 #import <UIKit/UIKit.h>
 
 @interface JCManChart : UIView
--(void)addLine:(int)endPointY;
+-(void)addLine:(float)endPointY;
 -(void)FinishTest;
 @end
